Revert "Don't use %ld with long variables on Windows."
This reverts commit 1c01fa35fa48761e14a1e0fd5e7af305aeb1af04.  Turned
out the actual problem was a bug in mingw's runtime library that only
occured after I upgraded my mingw installation. See also
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=452977

Conflicts:

	ChangeLog

+++ src/mmg/tabs/input.cpp
@@ -544,8 +544,8 @@ tab_input::add_file(const wxString &file_name,
 
         file->attached_files.push_back(a);
 
-        wxLogMessage(wxT("Attached file ID %d MIME type '%s' size %d description '%s' name '%s'"),
-                     int(a->id), a->mime_type.c_str(), int(a->size), a->description.c_str(), a->name.c_str());
+        wxLogMessage(wxT("Attached file ID %ld MIME type '%s' size %ld description '%s' name '%s'"),
+                     a->id, a->mime_type.c_str(), a->size, a->description.c_str(), a->name.c_str());
       }
 
     } else if (output[i].Find(wxT("Chapters")) == 0) {
@@ -1048,12 +1048,12 @@ tab_input::load(wxConfigBase *cfg,
     cfg->Read(wxT("container"), &fi->container);
     cfg->Read(wxT("appending"), &fi->appending, false);
 
-    int tidx;
+    long tidx;
     for (tidx = 0; tidx < (long)num_tracks; tidx++) {
       mmg_track_cptr tr(new mmg_track_t);
       bool dummy = false;
 
-      s.Printf(wxT("track %d"), tidx);
+      s.Printf(wxT("track %ld"), tidx);
 
       cfg->SetPath(s);
       wxString id;
